Image Settings update & Addition to Archetypes
Set the image configuration to not optimize so that when the user clicks on the image, it will enlarge it in the window. Since we are using webp files, we don't need to shrink the images further.

Archetypes, updated some settings that would be helpful to have on a new page.