Disabled Tags and Categories
While it is nice to have Tags and Categories, they are are not required to be displayed to the public. I've disabled them for the time being, but we will still use them in the background for SEO purposes. I've confirmed the tags will still display as keywords for the Header Tag for SEO when disabled.